+++ src/libs/notion/blocks/convert/childPage.ts
@@ -1,6 +1,6 @@
 import { isFullPage } from "@notionhq/client";
 
-import { retrievePage } from "../../pages/pages.js";
+import { retrievePage } from "../../pages/retrieve.js";
 
 import type { ChildPageBlockObject } from "../../../../types/notion/blocks/childPage.js";
 import type { ChildPageBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints.js";
