+++ src/libs/notion/blocks/convert/linkToPage.ts
@@ -2,7 +2,7 @@ import { isFullDatabase, isFullPage } from "@notionhq/client";
 
 import { listComments } from "../../comments.js";
 import { retrieveDatabase } from "../../database/retrieve.js";
-import { retrievePage } from "../../pages/pages.js";
+import { retrievePage } from "../../pages/retrieve.js";
 
 import type { LinkToPageBlockObject } from "../../../../types/notion/blocks/linkToPage.js";
 import type { LinkToPageBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints.js";
