+++ src/libs/notion/index.ts
@@ -1,43 +1,7 @@
-export * from "./blocks/convert/audio.js";
-export * from "./blocks/retrieve.js";
-export * from "./blocks/convert/bookmark.js";
-export * from "./blocks/convert/breadcrumb.js";
-export * from "./blocks/convert/bulletedListItem.js";
-export * from "./blocks/convert/callout.js";
-export * from "./blocks/convert/childDatabase.js";
-export * from "./blocks/convert/childPage.js";
-export * from "./blocks/convert/code.js";
-export * from "./blocks/convert/column.js";
-export * from "./blocks/convert/columnList.js";
-export * from "./blocks/convert/divider.js";
-export * from "./blocks/convert/embed.js";
-export * from "./blocks/convert/equation.js";
-export * from "./blocks/convert/file.js";
-export * from "./blocks/convert/heading1.js";
-export * from "./blocks/convert/heading2.js";
-export * from "./blocks/convert/heading3.js";
-export * from "./blocks/convert/image.js";
-export * from "./blocks/convert/linkPreview.js";
-export * from "./blocks/convert/linkToPage.js";
-export * from "./blocks/convert/numberedListItem.js";
-export * from "./blocks/convert/paragraph.js";
-export * from "./blocks/convert/pdf.js";
-export * from "./blocks/convert/quote.js";
-export * from "./blocks/convert/syncedBlock.js";
-export * from "./blocks/convert/tableRow.js";
-export * from "./blocks/convert/table.js";
-export * from "./blocks/convert/tableOfContents.js";
-export * from "./blocks/convert/template.js";
-export * from "./blocks/convert/toDo.js";
-export * from "./blocks/convert/toggle.js";
-export * from "./blocks/convert/unsupported.js";
-export * from "./blocks/convert/video.js";
-export * from "./richText/equation.js";
-export * from "./richText/mention.js";
-export * from "./richText/richText.js";
-export * from "./richText/richTextItem.js";
-export * from "./richText/text.js";
-export * from "./database/retrieve.js";
-export * from "./pages/pages.js";
+export * from "./blocks/index.js";
+export * from "./richText/index.js";
+export * from "./database/index.js";
+export * from "./pages/index.js";
+export * from "./properties/index.js";
 export * from "./parent.js";
 export * from "./comments.js";
