+++ src/libs/notion/pages/fetch.ts
@@ -1,7 +1,7 @@
 import { isFullPage } from "@notionhq/client";
 
 import { convertResponseToPage } from "./convert.js";
-import { retrievePage } from "./pages.js";
+import { retrievePage } from "./retrieve.js";
 
 import type { PageObject } from "../../../types/notion/pages/page.js";
 import type { Result } from "../../../types/utils.js";
