+++ src/libs/notion/parent.ts
@@ -2,7 +2,7 @@ import { isFullBlock, isFullPage, isFullDatabase } from "@notionhq/client";
 
 import { retrieveBlock } from "./blocks/retrieve.js";
 import { retrieveDatabase } from "./database/retrieve.js";
-import { retrievePage } from "./pages/pages.js";
+import { retrievePage } from "./pages/retrieve.js";
 
 import type {
   BlockObjectResponse,
